+++ src/core/reopt_multinode.jl
@@ -119,11 +119,6 @@ function add_variables!(m::JuMP.AbstractModel, ps::Array{REoptInputs})
                 sum( p.etariff.export_rates[:NEM][ts] * m[Symbol("dvNEMexport"*_n)][t, ts] for t in p.techs)
               + sum( p.etariff.export_rates[:WHL][ts] * m[Symbol("dvWHLexport"*_n)][t, ts]  for t in p.techs)
                 for ts in p.time_steps )
-            if !isempty(p.storage.export_bins)
-                m[Symbol("TotalExportBenefit"*_n)] += p.pwf_e * p.hours_per_timestep *
-                sum( p.etariff.export_rates[u][ts] * m[Symbol("dvStorageExport"*_n)][b,u,ts] 
-                     for b in p.storage.can_grid_charge, u in p.storage.export_bins)
-            end
         else
             m[Symbol("TotalExportBenefit"*_n)] = 0
         end
@@ -243,8 +238,6 @@ function build_reopt!(m::JuMP.AbstractModel, ps::Array{REoptInputs})
                             m[Symbol("dvProductionToStorage"*_n)][b, t, ts] == 0)
                 @constraint(m, [ts in p.time_steps], m[Symbol("dvDischargeFromStorage"*_n)][b, ts] == 0)
                 @constraint(m, [ts in p.time_steps], m[Symbol("dvGridToStorage"*_n)][b, ts] == 0)
-                @constraint(m, [u in p.storage.export_bins, ts in p.time_steps],
-                            m[Symbol("dvStorageExport"*_n)][b, u, ts] == 0)
             else
                 add_storage_size_constraints(m, p, b; _n=_n)
                 add_storage_dispatch_constraints(m, p, b; _n=_n)